create store and edit file in module and import them
