+++ src/modules/store.js
@@ -1,11 +1,5 @@
-import initList from './initList.js';
-
-export default () => {
-  const listName = initList();
-
-  const store = () => {
-    localStorage.setItem('tasks', JSON.stringify(listName));
-  };
-
-  return store;
+const store = (listName) => {
+  localStorage.setItem('tasks', JSON.stringify(listName));
 };
+
+export default store;
